Enabled the python tools to work with a non-zero snapshot
For this feature to be enabled a TreeFrog catalog of version>=1.21 is required as it uses the header information to know the snapshot offset that TreeFrog was run with. Otherwise the offset is set to zero. This maybe updated in the future to allow a user input snapshot offset.

The python tools now uses a dictionary to store the halodata and treedata so eliminating any index errors that might occur with lists. All the functions have been updated to include this offset