Bug fix to GenerateProgenitorLinks function
This code would run into an error if all haloes in a snapshot have no progenitors, this has now been fixed. In addition, the code has also been sped up through the use of (list).extend so concatination is not required and use of np.full to create any array of the same number.